Implement a generic RAII wrapper based on P0052R2
This is a generic scope exit wrapper that takes a callable and calls it
while the wrapper is destructed, unless the scope_exit is released
before it is destructed.

Use cases are for example general resource cleanup (close db
connections, file handles or similar), logging on exit or updating state
for the rest of the program if there are multiple points of return in a
function.